make image upload in to a importable function
